Danny's additions/changes/error handling
Added code to open "morse.txt" file, with an error msg if unsuccessful. Also added the appropriate error handling where indicated.One suggested change was on the letter_to_morse function, changed the "catch(int n)" to catch(...) because the compiler was complaining that the "ex" was an unreferenced local variable. Not sure how else to fix this, but the catch-all should work.